move the tracer-specific logic into a well-marked module and don't run code in global time if not lambda
