+++ .secrets.baseline
@@ -115,14 +115,14 @@
         "filename": "src/lumigo_tracer/lumigo_utils.py",
         "hashed_secret": "116b0dcbb1dfb3f19c902ebfc29f35d908cec6a4",
         "is_verified": false,
-        "line_number": 69
+        "line_number": 48
       },
       {
         "type": "Secret Keyword",
         "filename": "src/lumigo_tracer/lumigo_utils.py",
         "hashed_secret": "2a93870f6e6a2158cb9631b349e75d8f65fd81af",
         "is_verified": false,
-        "line_number": 70
+        "line_number": 49
       }
     ],
     "src/test/unit/event/test_event_dumper.py": [
@@ -184,7 +184,7 @@
         "filename": "src/test/unit/test_lumigo_utils.py",
         "hashed_secret": "f32b67c7e26342af42efabc674d441dca0a281c5",
         "is_verified": false,
-        "line_number": 263
+        "line_number": 199
       }
     ],
     "src/test/unit/test_tracer.py": [
@@ -220,5 +220,5 @@
       }
     ]
   },
-  "generated_at": "2022-11-08T16:26:37Z"
+  "generated_at": "2023-01-22T08:40:16Z"
 }
