+++ src/lumigo_tracer/__init__.py
@@ -1,4 +1,4 @@
-from .tracer import lumigo_tracer, LumigoChalice  # noqa
+from lumigo_tracer.lambda_tracer.tracer import lumigo_tracer, LumigoChalice  # noqa
 from .user_utils import (  # noqa
     report_error,
     add_execution_tag,
@@ -9,3 +9,6 @@ from .user_utils import (  # noqa
     error,
 )
 from .auto_instrument_handler import _handler  # noqa
+from .lambda_tracer.global_scope_exec import global_scope_exec
+
+global_scope_exec()
