+++ src/lumigo_tracer/user_utils.py
@@ -3,7 +3,7 @@ import logging
 from contextlib import contextmanager
 from typing import Dict, Optional
 
-from lumigo_tracer.spans_container import SpansContainer
+from lumigo_tracer.lambda_tracer.spans_container import SpansContainer
 from lumigo_tracer.lumigo_utils import warn_client, is_lambda_traced
 
 LUMIGO_REPORT_ERROR_STRING = "[LUMIGO_LOG]"
