+++ src/lumigo_tracer/wrappers/__init__.py
@@ -3,12 +3,14 @@ from .pymongo.pymongo_wrapper import wrap_pymongo
 from .redis.redis_wrapper import wrap_redis
 from .sql.sqlalchemy_wrapper import wrap_sqlalchemy
 from .aiohttp.aiohttp_wrapper import wrap_aiohttp
-
+from ..lumigo_utils import is_aws_environment
 
 already_wrapped = False
 
 
-def wrap(force: bool = False):  # type: ignore[no-untyped-def]
+def wrap(force: bool = False) -> None:
+    if not is_aws_environment():
+        return
     global already_wrapped
     if not already_wrapped:
         # Never wrap http calls twice - it will create duplicate body
