+++ src/lumigo_tracer/wrappers/aiohttp/aiohttp_wrapper.py
@@ -1,6 +1,6 @@
 from lumigo_tracer.lumigo_utils import lumigo_safe_execute, get_logger, concat_old_body_to_new
 from lumigo_tracer.libs.wrapt import wrap_function_wrapper
-from lumigo_tracer.spans_container import SpansContainer
+from lumigo_tracer.lambda_tracer.spans_container import SpansContainer
 from lumigo_tracer.wrappers.http.http_data_classes import HttpRequest
 from lumigo_tracer.wrappers.http.sync_http_wrappers import add_request_event, update_event_response
