+++ src/lumigo_tracer/wrappers/http/sync_http_wrappers.py
@@ -7,11 +7,12 @@ import http.client
 import logging
 import random
 
+from lumigo_tracer.lambda_tracer.lambda_reporter import get_edge_host
 from lumigo_tracer.wrappers.http.http_data_classes import HttpRequest, HttpState
 from lumigo_tracer.parsing_utils import safe_get_list, recursive_json_join
 from lumigo_tracer.wrappers.http.http_parser import get_parser, HTTP_TYPE
 from lumigo_tracer.libs.wrapt import wrap_function_wrapper
-from lumigo_tracer.spans_container import SpansContainer
+from lumigo_tracer.lambda_tracer.spans_container import SpansContainer
 from lumigo_tracer.lumigo_utils import (
     get_logger,
     lumigo_safe_execute,
@@ -20,7 +21,6 @@ from lumigo_tracer.lumigo_utils import (
     lumigo_dumps,
     get_size_upper_bound,
     is_error_code,
-    get_edge_host,
     TRUNCATE_SUFFIX,
     concat_old_body_to_new,
     EDGE_SUFFIX,
