+++ src/lumigo_tracer/wrappers/pymongo/pymongo_wrapper.py
@@ -7,7 +7,7 @@ from lumigo_tracer.lumigo_utils import (
     lumigo_dumps,
     get_current_ms_time,
 )
-from lumigo_tracer.spans_container import SpansContainer
+from lumigo_tracer.lambda_tracer.spans_container import SpansContainer
 
 try:
     from pymongo import monitoring
