+++ src/test/unit/lambda_tracer/test_spans_container.py
@@ -9,9 +9,11 @@ from datetime import datetime
 
 import pytest
 
-from lumigo_tracer import lumigo_utils, add_execution_tag
+from lumigo_tracer import add_execution_tag
+from lumigo_tracer.lambda_tracer import lambda_reporter
+from lumigo_tracer.lambda_tracer.lambda_reporter import get_extension_dir
 from lumigo_tracer.wrappers.http.http_parser import HTTP_TYPE
-from lumigo_tracer.spans_container import (
+from lumigo_tracer.lambda_tracer.spans_container import (
     SpansContainer,
     TimeoutMechanism,
     FUNCTION_TYPE,
@@ -53,7 +55,7 @@ def test_spans_container_not_send_start_span_on_send_only_on_errors_mode(monkeyp
 def test_start(monkeypatch):
     lumigo_utils_mock = mock.Mock()
     monkeypatch.setenv("LUMIGO_USE_TRACER_EXTENSION", "true")
-    monkeypatch.setattr(lumigo_utils, "write_extension_file", lumigo_utils_mock)
+    monkeypatch.setattr(lambda_reporter, "write_extension_file", lumigo_utils_mock)
     monkeypatch.setattr(SpansContainer, "_generate_start_span", lambda *args, **kwargs: {"a": "a"})
     monkeypatch.setattr(Configuration, "should_report", True)
     SpansContainer().start()
@@ -95,7 +97,7 @@ def only_if_error(dummy_span, monkeypatch, tmpdir):
 
     SpansContainer.get_span().add_span(dummy_span)
     reported_ttl = SpansContainer.get_span().end({})
-    stop_path_path = f"{lumigo_utils.get_extension_dir()}/span_name_stop"
+    stop_path_path = f"{get_extension_dir()}/span_name_stop"
     return reported_ttl, stop_path_path
