+++ src/test/unit/wrappers/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from lumigo_tracer.wrappers import wrap
+
+
+@pytest.fixture(autouse=True)
+def wrap_everything(aws_environment):
+    wrap()
