+++ src/test/unit/wrappers/sql/test_sqlalchemy_wrapper.py
@@ -7,8 +7,8 @@ from sqlalchemy import create_engine, Table, Column, Integer, String, MetaData
 from sqlalchemy.sql import select
 
 from lumigo_tracer.lumigo_utils import DEFAULT_MAX_ENTRY_SIZE
-from lumigo_tracer.spans_container import SpansContainer
-from lumigo_tracer.tracer import lumigo_tracer
+from lumigo_tracer.lambda_tracer.spans_container import SpansContainer
+from lumigo_tracer.lambda_tracer.tracer import lumigo_tracer
 
 TOKEN = "t_10faa5e13e7844aaa1234"
